+++ sixtyfps_runtime/interpreter/dynamic_component.rs
@@ -1195,7 +1195,7 @@ extern "C" fn layout_info(component: ComponentRefPin) -> LayoutInfo {
     // This is fine since we can only be called with a component that with our vtable which is a ComponentDescription
     let instance_ref = unsafe { InstanceRef::from_pin_ref(component, guard) };
 
-    let result = crate::eval_layout::get_layout_info(
+    let mut result = crate::eval_layout::get_layout_info(
         &instance_ref.component_type.original.root_element,
         instance_ref,
         &eval::window_ref(instance_ref).unwrap(),
@@ -1203,9 +1203,8 @@ extern "C" fn layout_info(component: ComponentRefPin) -> LayoutInfo {
 
     let constraints = instance_ref.component_type.original.root_constraints.borrow();
     if constraints.has_explicit_restrictions() {
-        let mut info = LayoutInfo::default();
         crate::eval_layout::fill_layout_info_constraints(
-            &mut info,
+            &mut result,
             &constraints,
             &|nr: &NamedReference| {
                 eval::load_property(instance_ref, &nr.element(), nr.name())
@@ -1214,10 +1213,8 @@ extern "C" fn layout_info(component: ComponentRefPin) -> LayoutInfo {
                     .unwrap()
             },
         );
-        result.merge(&info)
-    } else {
-        result
     }
+    result
 }
 
 unsafe extern "C" fn get_item_ref(component: ComponentRefPin, index: usize) -> Pin<ItemRef> {